+++ CHANGELOG.md
@@ -69,6 +69,8 @@ New Features
   PhaseScreenList. (#549)
 - Rewrote OpticalPSF to operate under the PhaseScreen framework to enable
   fully self-consistent optics+atmospheric PSFs. (#549)
+- Added a simple, linear model for persistence in the detectors that accepts a
+  list of galsim.Image instances and a list of an equal number of floats. (#554)
 - Added BoundsI.numpyShape() to easily get the numpy shape that corresponds
   to a given bounds instance. (#654)
 - Have FITS files with unsigned integer data automatically convert that into